Split `quad::Pipeline` into `core` and `compatibility`
