+++ glow/src/shader/compatibility/triangle.frag
@@ -1,0 +1,8 @@
+#version 100
+precision mediump float;
+
+varying vec4 v_Color;
+
+void main() {
+    gl_FragColor = v_Color;
+}
